fix(utils): fix routes v1/utils/version
Changelog
* Move function to get version from pyproject file to utils package
* Refactor main and utils_router to use this function
* Add test
