+++ boaviztapi/main.py
@@ -24,23 +24,10 @@ from boaviztapi.routers.server_router import server_router
 from boaviztapi.routers.cloud_router import cloud_router
 from boaviztapi.routers.terminal_router import terminal_router
 from boaviztapi.routers.utils_router import utils_router
+from boaviztapi.utils.get_version import get_version_from_pyproject
 
 from fastapi.responses import HTMLResponse
 
-def get_version_from_pyproject():
-    # List of potential locations for the pyproject.toml file
-    potential_paths = [
-        os.path.join(os.path.dirname(__file__), '../pyproject.toml'),
-        os.path.join(os.path.dirname(__file__), 'pyproject.toml'),
-    ]
-
-    for path in potential_paths:
-        if os.path.exists(path):
-            with open(path, 'r') as f:
-                return toml.loads(f.read())['tool']['poetry']['version']
-
-    # Raise an error if the file is not found in any of the locations
-    raise FileNotFoundError("pyproject.toml not found in expected locations")
 
 # Serverless frameworks adds a 'stage' prefix to the route used to serve applications
 # We have to manage it to expose openapi doc on aws and generate proper links.
